refactor: join train related utils
Move `getTrainType`, `sortSimpliedTrains`, `simplifyTrain` and `simplifyTrains` into a single file.
