+++ site/src/hooks/use_live_trains.ts
@@ -4,7 +4,7 @@ import type { SimplifiedTrain } from '@typings/simplified_train'
 import { useQuery } from '@tanstack/react-query'
 import { fetchLiveTrains } from '@junat/digitraffic'
 
-import { simplifyTrains } from '@utils/simplify_train'
+import { simplifyTrains } from '@utils/train'
 import { DEFAULT_TRAINS_COUNT, TRAINS_MULTIPLIER } from 'src/constants'
 
 type FetchDigitrafficProps = {
